channel#basic_cancel Subscriber should still process the messages it's assigned to
"This method cancels a consumer. This does not affect already delivered messages, but it does mean the server will not send any more messages for that consumer. The client may receive an arbitrary number of messages in between sending the cancel method and receiving the cancel-ok reply." - https://www.rabbitmq.com/amqp-0-9-1-reference.html#basic.cancel

Co-authored-by: Anders Bälter <anders@84codes.com>