Don't flush socket after publish if in transaction mode
Becuase that will cause a "write-write-read" situation, which without
tcp_nodelay will cause a long wait before the last read (Tx#CommitOk),
because of the delayed ack from the server, dropping the throughput substantially.
It's "safe" to buffer the publishes in the outgoing buffer as the
messages won't be actually published into the broker until the Tx#Commit
frame anyway.