Callbacks for blocked/unblocked notifications
When the server is out of disk or RAM it might notify its clients.
Prevoiusly we blocked all writes but that also made it impossible to
send Ack/Reject frames when consuming for instance, which made so that
the server might not recoup disk space or RAM when the queue was
shrinking.

Connection#on_blocked and Connection#on_unblocked callbacks are now
added so that the user can stop only publishing when the server is
blocked.